# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: MH Save Sync API
  version: 0.1.0-alpha
  summary: E2EE save snapshot coordination API; server stores opaque graph metadata and encrypted objects only.
servers:
  - url: http://127.0.0.1:18080
paths:
...
  /v1/snapshots/{upload_id}/chunks:
    post:
      summary: Upload one encrypted chunk with checksum.
      parameters:
        - name: upload_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Chunk durable }
  /v1/snapshots/{upload_id}/manifest:
    post:
      summary: Upload encrypted manifest after chunks.
      parameters:
        - name: upload_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Manifest durable }
  /v1/snapshots/{upload_id}/commit:
    post:
      summary: Insert immutable snapshot row and CAS the logical HEAD.
      description: If base_head is stale, the snapshot is retained as a conflict branch and HEAD is not overwritten.
      parameters:
        - name: upload_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Fast-forward, first snapshot or conflict outcome }
        '409': { description: Referenced object missing or session invalid }
...